libgav1: update snapshot to cl/263885228
+cl/264669395
significantly improves performance, this version now outperforms the
libaom based decoder.

Bug: 130249450
Test: aosp_arm-eng aosp_arm64-eng aosp_x86-eng aosp_x86_64-eng build
Test: cts-tradefed run commandAndExit cts-dev --include-filter CtsMediaTestCases --module-arg "CtsMediaTestCases:include-filter:android.media.cts.DecoderTest#testAV1*" --module-arg "CtsMediaTestCases:include-filter:android.media.cts.AdaptivePlaybackTest#testAV1*" (bonito-userdebug)
Test: (externally) libaom/Argon/Allegro test vectors, cpu/memory performance, fuzzing

+++ libgav1/src/quantizer.h
@@ -10,7 +10,8 @@ namespace libgav1 {
 
 // Stores the quantization parameters of Section 5.9.12.
 struct QuantizerParameters {
-  int16_t base_index;
+  // base_index is in the range [0, 255].
+  uint8_t base_index;
   int8_t delta_dc[kMaxPlanes];
   // delta_ac[kPlaneY] is always 0.
   int8_t delta_ac[kMaxPlanes];
